backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
python-dotenv
python-multipart
